1.5 rozbicie metody main, utworzenie klasy wewnętrznej Node
