+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -45,4 +45,36 @@ public class IntLinkedList {
 	public void setLast(Node last) {
 		this.last = last;
 	}
+
+	private class Node {
+
+		private final int value;
+		private Node prev, next;
+
+		public Node(int i) {
+			value = i;
+		}
+
+		public int getValue() {
+			return value;
+		}
+
+		public Node getPrev() {
+			return prev;
+		}
+
+		public void setPrev(Node prev) {
+			this.prev = prev;
+		}
+
+		public Node getNext() {
+			return next;
+		}
+
+		public void setNext(Node next) {
+			this.next = next;
+		}
+	}
 }
+
+
